bekrin-school1/bekrin-back/requirements.txt
Django
Pillow
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
drf-spectacular
django-environ
openpyxl
cryptography
